django-back/requirements.txt
Django==5.1.6
Django==5.1.7
djangorestframework==3.15.2
openpyxl==3.1.5
pandas==2.2.3
